+++ src/hipay_enterprise/classes/helper/HipayConfig.php
@@ -370,6 +370,7 @@ class HipayConfig
                     'card_token' => 0,
                     'activate_basket' => 1,
                     'log_infos' => 1,
+                    'log_debug' => 0,
                     'regenerate_cart_on_decline' => 1,
                     'ccDisplayName' => [
                         'fr' => 'Carte de crédit',
@@ -378,6 +379,7 @@ class HipayConfig
                     ],
                     'ccFrontPosition' => 1,
                     'send_url_notification' => 0,
+                    'display_cancel_button' => 0,
                 ],
                 'credit_card' => [],
                 'local_payment' => [],