+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.21.0
+
+- **Add**: Added reference to pay on customer order details for `Mooney` / `Multibanco` payment means
+- **Fix**: Fixed partial capture for `Alma` methods
+- **Add**: Display cancel button
+- **Add**: Added debug log option
+
 ## 2.20.0
 
 - **Add**: Added support for new payment means: